label_properties now generates a source location if there is none
A source_locationt with id() nil signals that the source location is not
given; an empty string signals that the source location is given.  Goto
instructions without source location are perfectly valid, but we need a
source location to carry the property id.  This fixes label_properties to
create a source location with the right function name if none is given.